Correccion data type, entidad
- se corrige data type de Gemini Servicio, de long a integer

-se pasa userId de int a long, debido a tipo solicitud jpa.

-se pasa long a integer en DocumentRepository, por motivo de uso jpa.

+++ Producto/Back-end/kydira-api/kydira-api/src/main/java/com/kydira_api/controller/SummaryController.java
@@ -21,7 +21,7 @@ public class SummaryController {
     private IGeminiService geminiService;
 
     @GetMapping("/generate/{documentId}")
-    public ResponseEntity<?> generateSummary(@PathVariable Long documentId) {
+    public ResponseEntity<?> generateSummary(@PathVariable Integer documentId) {
         try {
             Document doc = documentRepository.findById(documentId)
                 .orElseThrow(() -> new RuntimeException("Documento no encontrado"));
